pyproject.toml
[tool.poetry]
name = 'tea'
version = '0.1.0-rc1'
packages = [{include = 'tea', from='src'}]
description = 'Manage Telos EVM nodes using Ansible'
authors = ['Guillermo Rodriguez <guillermo@telos.net>']
license = 'MIT'
readme = 'README.md'

[tool.poetry.dependencies]
python = '>=3.10,<4'

rlp = '3.0.0'
requests = '<2.32.0'
ansible = '^9.6.0'
ansible-core = '<2.17.0'
click = '^8.1.7'
web3 = '^6.19.0'
bs4 = {version = '^0.0.1'}
zstandard = {version = '^0.21.0'}
py-leap = {git = 'https://github.com/guilledk/py-leap.git', rev = 'v0.1a22'}

[tool.poetry.group.dev.dependencies]
pdbpp = '^0.10.3'

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'
